fix(mcp): provide openSourceFile on non-unix platforms (windows build)
go vet on windows-latest failed at tools.go:1687 (undefined openSourceFile):
read_source_unix.go is tagged //go:build darwin || linux and read_source_windows.go
(//go:build windows) defined only readSourceWindow, leaving openSourceFile undefined
on Windows. BSDs had neither file.

Replace the windows-only read_source_windows.go with read_source_other.go tagged
//go:build !darwin && !linux, defining both openSourceFile (plain os.Open with the
Layer-1 regular-file parity check) and readSourceWindow. This covers windows + the
BSDs + any future GOOS. The darwin||linux file is unchanged (keeps the fsevents
O_NONBLOCK defense, getSourceSem, errGetSourceStall). Every GOOS now compiles
exactly one openSourceFile and one readSourceWindow.
